Create express app at server.js
I will need to query a third-party API, query from an internal database,
and populate a template with data upon receiving a request for the main
page. Express is a useful middleware library to intercept HTTP requests
and pass them along a chain of responsibility.

This commit creates a temporary dummy template for sanity testing.