Deploy to heroku; remove config.js from .gitignore
Deployment is trivial, and the project is in a functional enough
state that I can use it on the go.

This commit creates a Heroku configuration Procfile, and removes
api/config.js from .gitignore. Previously, this project was meant for
client-side rendering so the API keys in the config.js file had to be
hidden. But now that I am rendering on the server-side, I no longer need
to hide the API key.