Create scraper for Planned Work pages
At time of writing, no API exists to get the MTA's Planned Work. There's
an API to get emergency Service Advisories, but nothing to get, say,
repairs for the upcoming Saturday. This commit creates a scraper script
to extract the data from a given webpage.

This commit installs htmlrequest2 for parsing. htmlrequest2 is a
low-level library that functions through callbacks like onopentag,
ontext, and onclosetag.