Fix Heroku scraping bug
Remote heroku branch threw an error whenever the Scheduler add-on ran
the scrapeToDatabase task. This is because htmlparser2 was excluded from
the package.json file. While htmlparser2 was installed in local
node_modules, it was not listed as a dependency so the remote heroku
branch did not install it.

This commit adds htmlparser2 as a dependency to package.json to fix it.