Fix <input> UI bug in bus config <fieldset>
Multiple <input> and <label> elements were getting displayed as parts of
the same line. This caused a UI bug where a <input> box would be
followed by a <label> for the following <input> box, and where the
<label> would get split over multiple lines. This commit refactors the
html so that <label> and <input> gets interperted as block elements
instead.

Fixes #15.